// packages/runtime-vapor/js-reactive/types.d.ts
export type ObjectLike = Record<PropertyKey, unknown>
export interface HookHost extends ObjectLike {
  __hooks?: unknown
}
export interface HookContainer extends ObjectLike {
  states: unknown[]
  index: number
  __forcedIndex?: number
  __idMap?: Map<unknown, unknown>
}
export interface HookFrame {
  instance: unknown
  hooks?: HookContainer
}
export interface HookCarrier {
  getCurrentInstance(): unknown
  renderHooks<T>(instance: unknown, render: () => T): T
}
export interface HookContext extends HookCarrier {
  setCurrentInstance(instance: unknown): void
  vaporWithHookId<T>(id: unknown, runner: () => T): T
  withHookSlot<T>(factory: () => T): T
}
export type ComputedGetter<T> = () => T
export interface ComputedOptions<T> {
  get: ComputedGetter<T>
  set?: (value: T) => void
}
export type ComputedInput<T> = ComputedGetter<T> | ComputedOptions<T>
export interface ReadableSignalHandle<T> extends KernelReadonlySignal<T>, ObjectLike {}
export interface ComputedHandle<T> extends KernelWritableSignal<T>, ObjectLike {
  __rueInvalidateComputed?: () => void
}
export interface ComputedHolder<T> {
  arg: ComputedInput<T>
}
export interface ComputedSlot<T> {
  handle: ComputedHandle<T>
  holder: ComputedHolder<T>
}
export type EffectScopeHandle = number
export interface CreateComputedHooksOptions {
  context: HookContext
  reactiveRuntime: unknown
  createComputed: unknown
}